requirements.txt
streamlit>=1.30
transformers==4.41.1
torch==2.3.0
llama-cpp-python==0.2.35
fer
opencv-python-headless
opensmile==4.0.1
librosa
soundfile
numpy
scikit-learn